OptionWindow clean up
The radio button generation code has been cleaned up and therefore also the way the background information is stored and resolved
Having a list of all backgrounds explicitly allows to generate the radios with a map over that list. It also gives the flexibility needed for background configurations that don't neatly fit into a x - x - x set.